assimilate userinfocontext with authstatecontext
