+++ app/src/components/activity/ActivityComponent.tsx
@@ -19,7 +19,7 @@ import { sanitizeRecord } from 'utils/addActivity';
 import { useKeycloak } from '@react-keycloak/web';
 import { calc_lat_long_from_utm } from 'components/map/Tools/DisplayPosition';
 import { Feature } from '@turf/turf';
-import { UserInfoContext } from 'contexts/UserInfoContext';
+import { AuthStateContext } from 'contexts/authStateContext';
 
 export interface IActivityComponentProps extends IMapContainerProps, IFormContainerProps, IPhotoContainerProps {
   classes?: any;
@@ -42,7 +42,7 @@ const ActivityComponent: React.FC<IActivityComponentProps> = (props) => {
   const databaseContext = useContext(DatabaseContext2);
   const dataAccess = useDataAccess();
   const { keycloak } = useKeycloak();
-  const { userInfo } = useContext(UserInfoContext);
+  const { userInfo } = useContext(AuthStateContext);
 
   const getLocation = async () => {
     const position = await Geolocation.getCurrentPosition();
